Added functionality for dynamic quizes
Quizes are now created more naturally, question by question, rather than as a combination of string. A question is created by a + button, the question is entered, and the question type of multi, float slider or int slider selected. With js this creates the body of the question allowing you to input options. Answers still to be implemented